+++ lib/vpn/vpn_switch.dart
@@ -17,21 +17,9 @@ class VPNSwitch extends StatefulWidget {
 //https://pub.dev/packages/animated_toggle_switch
 class _VPNSwitchState extends State<VPNSwitch> {
   final adHelper = AdHelper();
-  String vpnStatus = 'disconnected';
 
   bool isIdle(String vpnStatus) =>
       vpnStatus != 'connecting' && vpnStatus != 'disconnecting';
-
-  Future<void> vpnProcessForDesktop() async {
-    bool isConnected = vpnStatus == 'connected';
-    if (isConnected) {
-      LanternFFI.sysProxyOff();
-      await TrayHandler.instance.setupTray(false);
-    } else {
-      LanternFFI.sysProxyOn();
-      await TrayHandler.instance.setupTray(true);
-    }
-  }
 
   Future<void> vpnProcessForMobile(
       bool newValue, String vpnStatus, bool userHasPermission) async {
@@ -64,6 +52,7 @@ class _VPNSwitchState extends State<VPNSwitch> {
   Widget build(BuildContext context) {
     final internetStatusProvider = context.watch<InternetStatusProvider>();
     final vpnNotifier = context.watch<VPNChangeNotifier>();
+
     if (isMobile()) {
       return sessionModel
           .shouldShowGoogleAds((context, isGoogleAdsEnable, child) {
@@ -98,33 +87,25 @@ class _VPNSwitchState extends State<VPNSwitch> {
     } else {
       // This ui for desktop
       return Transform.scale(
-        scale: 2.5,
-        child: vpnModel
-            .vpnStatus((BuildContext context, String vpnStatus, Widget? child) {
-          this.vpnStatus = vpnStatus;
-          return AdvancedSwitch(
+          scale: 2.5,
+          child: AdvancedSwitch(
             width: 60,
             disabledOpacity: 1,
             enabled: (internetStatusProvider.isConnected &&
                 !vpnNotifier.isFlashlightInitializedFailed),
-            initialValue: this.vpnStatus == 'connected' ||
-                this.vpnStatus == 'disconnecting',
+            initialValue: vpnNotifier.vpnStatus.value == 'connected' ||
+                vpnNotifier.vpnStatus.value == 'disconnecting',
             activeColor: onSwitchColor,
             inactiveColor: (internetStatusProvider.isConnected &&
                     !vpnNotifier.isFlashlightInitializedFailed)
                 ? offSwitchColor
                 : grey3,
             onChanged: (newValue) {
-              vpnProcessForDesktop();
               WidgetsBinding.instance.addPostFrameCallback((_) {
-                setState(() {
-                  this.vpnStatus = newValue ? 'connected' : 'disconnected';
-                });
+                vpnNotifier.toggleConnection();
               });
             },
-          );
-        }),
-      );
+          ));
     }
   }
 }
